% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyAPI.R
\name{anyAPI}
\alias{anyAPI}
\title{(Advanced) A generalized function for interacting directly with any Data Library App API.}
\usage{
anyAPI(
  appName = c("magma", "janus", "metis", "timur", "vulcan", "polyphemus", "vesta",
    "etna"),
  api,
  target,
  method = c("get", "post", "delete"),
  requestBody = NULL,
  raw = FALSE,
  jsonparse = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{appName}{String providing the name of the app with the API you wish to target}

\item{api}{String providing the endpoint to hit.  e.g. "retrieve" or "gnomon/<project>/rules"}

\item{target}{A list, which can be created using \code{\link{magmaRset}}, containing your authorization 'token' (a string), a 'url' of magma to target (a string), and optional 'opts' for specifying additions parameters for curl requests (a named list).}

\item{method}{"get", "post", or "delete". String providing the http protocol used for this API}

\item{requestBody}{JSON-format String OR a named list providing any additional parameters used by the given API.
When given as a list, where names are the parameter names and contents are the intended values, the contents will be converted to a JSON string internally.}

\item{jsonparse}{Logical setting whether the return should be converted from JSON-string format into an analogous R format}
}
\value{
Default = Any structure determined by the behavior of the given API, OR a crul response object if \code{raw} is set to \code{TRUE}.
}
\description{
(Advanced) A generalized function for interacting directly with any Data Library App API.
}
\examples{
if (interactive()) {
    # First, we use magmaRset to create an object which will tell other magmaR
    #  functions our authentication token (as well as some other optional bits).
    # When run in this way, it will ask you to give your token.
    magma <- magmaRset()

    # Retrieving gnomon rules of a project via the gnomon#rules API
    rules_raw <- anyAPI("magma", paste0("gnomon/example/rules"), prod, 'get')
    rules <- rules_raw$rules
    print(rules)
}
}
