% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-adjust.R
\name{authentication-and-environments}
\alias{authentication-and-environments}
\title{Starting out and working with different Etna environments}
\description{
Starting out and working with different Etna environments
}
\section{Authorization via 'token's}{

Access to magma via magmaR is authenticated via 'token's which users can obtain from Janus.
A valid token must be provided to magmaR before any calls to magma can be successfully performed.
To provide this token, users should obtain their token from Janus, then provide it to magmaR functions using the \code{\link{magmaRset}} function.
See the function's own documentation and other functions' examples for further details and usage code.
}

\section{Non-production Environments}{

The Mount Etna code base relies on 3 different ecosystems of all of its components for purposes of
"development" of new tools and features,
"staging" of updates and data prior to release, and 
"production", the release version which most users see.
We won't get into the details too much more here, but users with access to the non-production environments can access these environments with magmaR.

To target non-production environments, users should provide the \code{url} of their alternative version of magma to magmaR functions using the \code{\link{magmaRset}} function.
If proxy or other curl-request settings need to be adjusted, users can provide these via the \code{opts} input of this same \code{\link{magmaRset}} function.
}

\seealso{
\code{\link{magmaRset}}
}
