% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveAttributes.R
\name{retrieveAttributeNames}
\alias{retrieveAttributeNames}
\title{Determine the attribute options associated with a given project-model pair.}
\usage{
retrieveAttributeNames(
  projectName = "ipi",
  modelName = "patient",
  recordNames = "all",
  token = .get_TOKEN(),
  ...
)
}
\arguments{
\item{projectName}{the name of the project you would like to download data from, exactly as it appears at https://timur.ucsf.edu}

\item{modelName}{the name of the piece of the magma model form which to download data}

\item{recordNames}{which records and attributes to grab}

\item{token}{your TOKEN from https://janus.ucsf.edu}

\item{...}{Additional parameters passed along to the internal `.retrieve()` function.
For troubleshooting or privileged-user purposes only.
Options: \code{request.only} (Logical), \code{json.params.only} (Logical), \code{verbose} (Logical), or \code{url.base} (String which can be user to direct toward production versus staging versus development versions of maagma).}
}
\value{
a string vector
}
\description{
Determine the attribute options associated with a given project-model pair.
}
\examples{

if (interactive()) {
    # Running like this will ask for input of your janus token.
    retrieveAttributeNames(
        projectName = "ipi",
        modelName = "patient")
}

}
