% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveAttributes.R
\name{retrieveAttributes}
\alias{retrieveAttributes}
\alias{retrieveAttributeNames}
\alias{retrieveAttributeCounts}
\title{Assess the downloadable data from magma}
\usage{
retrieveAttributeNames(
  projectName = "ipi",
  modelName = "patient",
  recordNames = "all",
  token = .get_TOKEN(),
  ...
)

retrieveAttributeCounts(
  projectName = "ipi",
  modelName = "patient",
  recordNames = "all",
  attributeNames = "all",
  token = .get_TOKEN(),
  ...
)
}
\arguments{
\item{projectName}{the name of the project you would like to download data from, exactly as it appears at https://timur.ucsf.edu}

\item{modelName}{the name of the piece of the magma model form which to download data}

\item{recordNames, attributeNames}{particular record(s) or attribute(s) to grab names for}

\item{token}{your TOKEN from https://janus.ucsf.edu}

\item{...}{Additional parameters passed along to the internal `.retrieve()` function.
For troubleshooting or privileged-user purposes only.
Options: \code{request.only} (Logical), \code{json.params.only} (Logical), \code{verbose} (Logical), or \code{url.base} (String which can be user to direct toward production versus staging versus development versions of maagma).}
}
\value{
a string vector
}
\description{
Assess the downloadable data from magma
}
\section{Functions}{
\itemize{
\item \code{retrieveAttributeNames}: Assess the names of attributes downloadable from magma

\item \code{retrieveAttributeCounts}: Assess the counts of records for attributes downloadable from magma
}}

\examples{

if (interactive()) {
    # Running like this will ask for input of your janus token.
    retrieveAttributeNames(
        projectName = "ipi",
        modelName = "patient")
}

}
