% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveMatrix.R
\name{retrieveMatrix}
\alias{retrieveMatrix}
\title{Download data stored as a matrix}
\usage{
retrieveMatrix(
  projectName,
  modelName,
  recordNames,
  attributeNames = "all",
  filter = "",
  page = NULL,
  pageSize = 10,
  token = .get_TOKEN(),
  ...
)
}
\arguments{
\item{projectName}{the name of the project you would like to download data from, exactly as it appears at https://timur.ucsf.edu}

\item{modelName}{the name of the piece of the magma model form which to download data}

\item{recordNames}{which records and attributes to grab}

\item{attributeNames}{which records and attributes to grab}

\item{filter}{potential filter of the data}

\item{page}{Integers. For retrieving just a portion of the data, sets slice/page size which is equivalent to a number of rows, and which slice to get.}

\item{pageSize}{Integers. For retrieving just a portion of the data, sets slice/page size which is equivalent to a number of rows, and which slice to get.}

\item{token}{your TOKEN from https://janus.ucsf.edu}

\item{...}{Additional parameters passed along to the internal `.retrieve()` function.
For troubleshooting or privileged-user purposes only.
Options: \code{request.only} (Logical), \code{json.params.only} (Logical), \code{verbose} (Logical), or \code{url.base} (String which can be user to direct toward production versus staging versus development versions of maagma).}
}
\value{
a matrix
}
\description{
Download data stored as a matrix
}
\examples{

# Unless a working TOKEN is hard-coded, or it is in an interactive session,
#   this code will not work.

if (interactive()) {
    # Running like this will ask for input of your janus token on first.
    
    ids <- retrieve("ipi", "rna_seq", attributeNames = "identifier")
    
    retrieveMatrix(
        projectName = "ipi",
        modelName = "rna_seq",
        recordNames = ids[1:4],
        attributeNames = "gene_counts")
}

}
