% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveMetadata.R
\name{retrieveMetadata}
\alias{retrieveMetadata}
\title{Retrieve data from one model of a project, metadata, for records linked to data from a different, target model of the same project}
\usage{
retrieveMetadata(
  projectName,
  meta_modelName,
  meta_attributeNames = "all",
  target_modelName,
  target_recordNames = "all",
  meta_group.by = NULL,
  token = .get_TOKEN(),
  ...
)
}
\arguments{
\item{projectName}{Single string. The name of the project you would like to download data from, exactly as it appears at https://timur.ucsf.edu}

\item{meta_modelName, meta_attributeNames}{Strings which indicate the linked "meta"data to retrieve.}

\item{target_modelName, target_recordNames}{Strings which indicate the 'target' data that meta-data is desired to be linked to.}

\item{meta_group.by}{\emph{Not implemented yet.} String, an attribute within the meta_model that can be used to group data when there are expected to be more than 1 record of metadata per record of target data.}

\item{token}{your personal TOKEN from \url{https://janus.ucsf.edu}. 

When not explicitly given in a function call: you will be prompted to input token, one time.
This user provided token will then be stored as a hidden variable, \code{.TOKEN}, in the global R environment,
and all future magmaR calls without an explicitly provided token will turn to this \code{.TOKEN}.}

\item{...}{Additional parameters passed along to the internal `.query()` and `.retrieve()` functions.
For troubleshooting or privileged-user purposes only.
Options: \code{verbose} (Logical), or \code{url.base} (String which can be user to direct toward production versus staging versus development versions of maagma).}
}
\value{
A dataframe with rows = target_recordNames and columns = either meta_attributeNames or repeats of meta_attributeNames to accommodate 1:many mappings of target data records to metadata records.
}
\description{
Retrieve data from one model of a project, metadata, for records linked to data from a different, target model of the same project
}
\examples{

if (interactive()) {
    # Running like this will ask for input of your janus token one time.
    retrieveMetadata(
        projectName = "ipi",
        meta_modelName = "sample",
        meta_attributeNames = "tissue_type",
        target_modelName = "rna_seq",
        filter = "")
}

}
