% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveProjects.R
\name{retrieveProjects}
\alias{retrieveProjects}
\title{Retrieve all the projectName options that a user has access to, from janus.}
\usage{
retrieveProjects(token = .get_TOKEN(), url.base = .get_URL(), verbose = FALSE)
}
\arguments{
\item{token}{Single string. Your personal token from \url{https://janus.ucsf.edu}. 

When not explicitly given in a function call: you will be prompted to input token, one time.
This user provided token will then be stored as a hidden variable, \code{.MAGMAR_TOKEN}, in the global R environment,
and all future magmaR calls without a \code{token} explicitly provided will turn to this \code{.MAGMAR_TOKEN}.}

\item{url.base}{Single string. A url which directs towards production versus staging versus development etna environments. See \code{\link{magma-environments}}.

Note that here, we actually pull from 'janus' rather than 'magma', but the function will correct any magma-targeting urls to target janus.}

\item{verbose}{Logical. Sets whether to report the status of the '/projects' curl request sent to janus.}
}
\value{
A data.frame where elements of the 'project_name' column reflect what can be given to \code{projectName} inputs of other magmaR functions.
}
\description{
Retrieve all the projectName options that a user has access to, from janus.
}
\details{
This function takes in a user's authorization token, and a url targetting either magma or janus.
It then converts that url to target janus, and makes a curl request to <janus-url>/projects in order to return which projects a user can access.
}
\examples{

if (interactive()) {

    # Running like this will ask for input of your janus token one time.
    retrieveProjects()
}

}
