% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateMagma}
\alias{updateMagma}
\title{Analogous to the '/update' function of magma}
\usage{
updateMagma(projectName, revisions = list(), token = .get_TOKEN(), ...)
}
\arguments{
\item{projectName}{Single string. The name of the project you would like to download data from, exactly as it appears at https://timur.ucsf.edu}

\item{revisions}{A list of named lists containing the data to be updated.

List structure:
\itemize{
\item top level name(s): modelNames, can be 1 or more.
\item 2nd level name(s): recordNames, can be 1 or more.
\item 3rd level name(s) & contents: the attributes to update & the values to use.
}

See \url{https://mountetna.github.io/magma.html#update} for additional formatting details.}

\item{token}{your personal TOKEN from \url{https://janus.ucsf.edu}. 

When not explicitly given in a function call: you will be prompted to input token, one time.
This user provided token will then be stored as a hidden variable, \code{.TOKEN}, in the global R environment,
and all future magmaR calls without an explicitly provided token will turn to this \code{.TOKEN}.}

\item{...}{Additional parameters passed along to the internal `.update()` function.
For troubleshooting or privileged-user purposes only.
Options: \code{request.only} (Logical), \code{json.params.only} (Logical), \code{verbose} (Logical), or \code{url.base} (String which can be used to direct toward production versus staging versus development versions of magma).}
}
\value{
None directly.  The function sends data to magma, and the only output is whether that send was successful.
}
\description{
Analogous to the '/update' function of magma
}
\details{
This function initially mimics the activity of the magma/query which is documented here \url{https://mountetna.github.io/magma.html#query}.

Afterwards, the json list output of magma/query is converted into an R list, and then the \code{format} input determines whether it should be wrangled further:
\itemize{
\item \code{format = "list"}, default: R list output directly.
\item \code{format = "df"}: R list converted into a dataframe where data comes from the list$answer and column names come from the list$format
}
}
\examples{

if (interactive()) {
    # Running like this will ask for input of your janus token one time.
    
    ### To obtain the sample-model record-identifiers that are linked to
    #   records of the rna_seq-model:

    # "Raw" output of query:
    query(
        projectName = projectName,
        queryTerms = 
            list('rna_seq',
                 '::all',
                 'sample',
                 '::identifier'))
                 
    # Or instead re-formatted to a dataframe, which may be easier for
    #   downstream applications in R:
    query(
        projectName = projectName,
        queryTerms = 
            list('rna_seq',
                 '::all',
                 'sample',
                 '::identifier'),
        format = 'df')
}

}
\seealso{
\url{https://mountetna.github.io/magma.html#update} for documentation of the underlying magma/update function.
}
