% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Analagous to the '/query' fucntion of magma}
\usage{
query(
  projectName,
  queryTerms = list(),
  format = c("list", "df"),
  token = .get_TOKEN(),
  ...
)
}
\arguments{
\item{projectName}{the name of the project you would like to download data from, exactly as it appears at https://timur.ucsf.edu}

\item{queryTerms}{An array of query predicates, see \url{https://mountetna.github.io/magma.html#query}}

\item{format}{Either "list" or "df" (=dataframe), this sets the desired output format}

\item{token}{your TOKEN from https://janus.ucsf.edu}

\item{...}{Additional parameters passed along to the internal `.query()` function.
For troubleshooting or privileged-user purposes only.
Options: \code{request.only} (Logical), \code{json.params.only} (Logical), \code{verbose} (Logical), or \code{url.base} (String which can be user to direct toward production versus staging versus development versions of maagma).}
}
\value{
Either list conversion of the raw '/query' JSON output if \code{format == "list"}, default,

OR a dataframe conversion if \code{format = "df"}
}
\description{
Analagous to the '/query' fucntion of magma
}
\examples{

##### FILLER

# Unless a working TOKEN is hard-coded, or it is in an interactive session,
#   this code will not work.

if (interactive()) {
    # Running like this will ask for input of your janus token.
    retrieveJSON(
        projectName = "ipi",
        modelName = "patient",
        recordNames = "all",
        attributeNames = "all",
        filter = "")
}

}
