% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\title{Download data from magma as a tsv and convert to a data.frame}
\usage{
retrieve(
  projectName,
  modelName,
  recordNames = "all",
  attributeNames = "all",
  filter = "",
  page = NULL,
  pageSize = 10,
  connected.only = TRUE,
  token = .get_TOKEN(),
  ...
)
}
\arguments{
\item{projectName}{the name of the project you would like to download data from, exactly as it appears at https://timur.ucsf.edu}

\item{modelName}{the name of the piece of the magma model form which to download data}

\item{recordNames, attributeNames}{which records and attributes to grab}

\item{filter}{potential filter of the data}

\item{pageSize, page}{Integers. For retrieving just a portion of the data, sets slice/page size which is equivalent to a number of rows, and which slice to get.}

\item{connected.only}{Logical. Whether data without a parent should be retained.}

\item{token}{your TOKEN from https://janus.ucsf.edu}

\item{...}{Additional parameters passed along to the internal `.retrieve()` function.
For troubleshooting or privileged-user purposes only.
Options: \code{request.only} (Logical), \code{json.params.only} (Logical), \code{verbose} (Logical), or \code{url.base} (String which can be user to direct toward production versus staging versus development versions of maagma).}
}
\value{
A dataframe representation of the requested data (or a list if the data was not rectangular).
}
\description{
Download data from magma as a tsv and convert to a data.frame
}
\examples{

# Unless a working TOKEN is hard-coded, or it is in an interactive session,
#   this code will not work.

if (interactive()) {
    # Running like this will ask for input of your janus token.
    retrieve(
        projectName = "ipi",
        modelName = "patient",
        recordNames = "all",
        attributeNames = "all",
        filter = "")
}

}
