% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-special.R
\name{retrieve_SpecialCases}
\alias{retrieve_SpecialCases}
\alias{retrieveTemplate}
\alias{retrieveModels}
\alias{retrieveIds}
\alias{retrieveAttributes}
\title{Special cases for the magma retrieve function, as laid out in https://mountetna.github.io/magma.html#retrieve}
\usage{
retrieveTemplate(projectName, token = .get_TOKEN(), ...)

retrieveModels(projectName, token = .get_TOKEN(), ...)

retrieveIds(projectName, modelName, token = .get_TOKEN(), ...)

retrieveAttributes(projectName, modelName, token = .get_TOKEN(), ...)
}
\arguments{
\item{projectName}{Single string. The name of the project you would like to download data from, exactly as it appears in the project_name column of https://timur.ucsf.edu}

\item{token}{Single string. Your personal token from \url{https://janus.ucsf.edu}. 

When not explicitly given in a function call: you will be prompted to input token, one time.
This user provided token will then be stored as a hidden variable, \code{.MAGMAR_TOKEN}, in the global R environment,
and all future magmaR calls without a \code{token} explicitly provided will turn to this \code{.MAGMAR_TOKEN}.}

\item{...}{Additional parameters passed along to the internal `.retrieve()` function,
for troubleshooting or advanced-user purposes only: \itemize{
\item \code{request.only} (Logical) & \code{json.params.only} (Logical) which stop the function short and return these values that would have been sent to magma.
\item \code{verbose} (Logical) sets whether to report the status of the '/retrieve' curl request sent to magma.
\item \code{url.base} (String) used to direct towards production versus staging versus development versions of magma. See \code{\link{magma-environments}}
}}

\item{modelName}{Single string. The name of data structure within your project, which are referred to as 'model's in magma, form which to download data.
For options, see \code{\link{retrieveModels}} or https://timur.ucsf.edu/<projectName>/map.}
}
\value{
retrieveTemplate = a list conversion of the project's template json.

retrieveModels = a string vector of model names

retrieveIds = a string vector of record names/identifiers.

retrieveAttributes = a string vector of attribute names.
}
\description{
Special cases for the magma retrieve function, as laid out in https://mountetna.github.io/magma.html#retrieve
}
\section{Functions}{
\itemize{
\item \code{retrieveTemplate}: Retrieve the template for a given project

\item \code{retrieveModels}: Retrieve the modelNames for a given project

\item \code{retrieveIds}: Retrieve all the identifiers/recordNames for a given project-model pair.

\item \code{retrieveAttributes}: Retrieve all the attribute options for a given project-model pair.
}}

\examples{

if (interactive()) {
    # Running like this will ask for input of your janus token one time.
    
    retrieveTemplate(
        projectName = "example")
        
    retrieveModels(
        projectName = "example")
        
    retrieveIds(
        projectName = "example",
        modelName = "rna_seq")
        
    retrieveAttributes(
        projectName = "example",
        modelName = "subject")
}
}
