% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-special.R
\name{retrieve_SpecialCases}
\alias{retrieve_SpecialCases}
\alias{retrieveTemplate}
\alias{retrieveIds}
\title{Special cases for the magma retrieve function as laid out in https://mountetna.github.io/magma.html#retrieve}
\usage{
retrieveTemplate(projectName, token = .get_TOKEN(), ...)

retrieveIds(projectName, modelName, token = .get_TOKEN(), ...)
}
\arguments{
\item{projectName}{the name of the project you would like to download data from, exactly as it appears at https://timur.ucsf.edu}

\item{token}{your TOKEN from https://janus.ucsf.edu}

\item{...}{Additional parameters passed along to the internal `.retrieve()` function.
For troubleshooting or privileged-user purposes only.
Options: \code{request.only} (Logical), \code{json.params.only} (Logical), \code{verbose} (Logical), or \code{url.base} (String which can be user to direct toward production versus staging versus development versions of maagma).}

\item{modelName}{the name of the piece of the magma model form which to download data}
}
\value{
retrieveTemplate = a jsonlite::fromJSON conversion of template JSON.

retrieveIds = a string vector
}
\description{
Special cases for the magma retrieve function as laid out in https://mountetna.github.io/magma.html#retrieve
}
\section{Functions}{
\itemize{
\item \code{retrieveTemplate}: Retrieve the template for a given project

\item \code{retrieveIds}: Retrieve all the identifiers for a given project-model pair.
}}

\examples{

if (interactive()) {
    # Running like this will ask for input of your janus token.
    
    retrieveTemplate(
        projectName = "ipi")
        
    retrieveIds(
        prjectName = "ipi",
        modelName = "rna_seq")
}
}
