% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveMatrix.R
\name{retrieveMatrix}
\alias{retrieveMatrix}
\title{Download data from magma that is stored as a matrix}
\usage{
retrieveMatrix(
  projectName,
  modelName,
  recordNames,
  attributeNames = "all",
  filter = "",
  page = NULL,
  pageSize = 10,
  token = .get_TOKEN(),
  ...
)
}
\arguments{
\item{projectName}{Single string. The name of the project you would like to download data from, exactly as it appears at https://timur.ucsf.edu}

\item{modelName}{Single string. The name of data structure within your project, which are referred to as 'model's in magma, form which to download data.
For options, see \code{\link{retrieveModels}} or https://timur.ucsf.edu/<projectName>/map.}

\item{recordNames}{Single string or string vector. Which particular sample/tube/etc. records to grab data for.
Options are "all" or any combination of individual record names. To retrieve individual options, see \code{\link{retrieveIds}}.}

\item{attributeNames}{Single string or string vector. Which features of the data to obtain.
Options are "all" or any combination of individual attribute names. To retrieve individual options, see \code{\link{retrieveAttributes}}.}

\item{filter}{String. Potential filter of the data.
Example: "<targetAttributeName>~GYN" to filter to records where <targetAttributeName> contains "GYN".

Refer to \url{https://mountetna.github.io/magma.html#retrieve} for more details.}

\item{page}{Integers. For retrieving just a portion of the data, sets slice/page size, which is equivalent to the a number of rows, and which slice to get.}

\item{pageSize}{Integers. For retrieving just a portion of the data, sets slice/page size, which is equivalent to the a number of rows, and which slice to get.}

\item{token}{your personal TOKEN from \url{https://janus.ucsf.edu}. 

When not explicitly given in a function call: you will be prompted to input token, one time.
This user provided token will then be stored as a hidden variable, \code{.TOKEN}, in the global R environment,
and all future magmaR calls without an explicitly provided token will turn to this \code{.TOKEN}.}

\item{...}{Additional parameters passed along to the internal `.retrieve()` function.
For troubleshooting or privileged-user purposes only.
Options: \code{request.only} (Logical), \code{json.params.only} (Logical), \code{verbose} (Logical), or \code{url.base} (String which can be used to direct toward production versus staging versus development versions of magma).}
}
\value{
a matrix
}
\description{
Download data from magma that is stored as a matrix
}
\examples{

if (interactive()) {
    # Running like this will ask for input of your janus token one time.
    
    ids <- retrieveIds("ipi", "rna_seq")
    
    retrieveMatrix(
        projectName = "ipi",
        modelName = "rna_seq",
        recordNames = ids[1:4],
        attributeNames = "gene_counts")
}

}
